multiagent/requirements.txt
langchain
langgraph
langchain-openai
pydantic
python-dotenv
redis
langgraph-checkpoint-redis
fastapi
uvicorn
fastmcp
langchain-mcp-adapters
httpx
